# repo-updater: Faster counting of user added repos